1. Moved the homeworks array to dashboard.vue where it is now passed to the list component via props. Additionaly adjusted some of the unit tests accordingly. 2. Changed index.vue to dashboard.vue and deleted 'dashboard' folder.
