+++ src/pages/beta/dashboard.vue
@@ -1,17 +1,23 @@
 <template>
 	<v-container class="v-container">
-		<h1 class="h4">{{ $t("pages.homeworks.title") }}</h1>
-		<homeworks-list />
+		<h1 class="h4" >{{ $t("pages.homeworks.title") }}</h1>
+		<homeworks-list :homeworks="homeworks"/>
 	</v-container>
 </template>
 
 <script>
 import HomeworksList from "@components/organisms/HomeworksList";
+import { mapGetters } from "vuex";
 
 export default {
 	components: { HomeworksList },
 	layout: "defaultVuetify",
-	created() {
+	computed: {
+		...mapGetters("homeworks", {
+			homeworks: "list",
+		}),
+	},
+	mounted() {
 		this.$store.dispatch("homeworks/getHomeworksDashboard");
 	},
 	head() {
