--- src/pages/beta/dashboard/index.unit.js
@@ -1,7 +1,0 @@
-describe("Homeworks/index", () => {
-	it.todo("Should render homeoworks list component");
-
-	it.todo("Should show a custom page title");
-
-	it.todo("Should should trigger a store action");
-});
